Incomplete Nether blocks, other tweaks Refactor wood lang keys to use the type_thing format instead of thing_type, I also edited existing lang files for this refactor (<3 kotmatross28729 and Omgise, that'll make things easier for you) Some related refactoring was needed for the easier introduction of new wood types. Signs and wood stuff was heavily refactored to not rely on vanilla planks. Signs do a similar thing to slabs now where they remember the previously registered sign so we don't have to pass each wall sign into the constructor. The flower item block has been renamed to potable item block because it could really just be any bush or cactus, not a flower. The sign packet is now a lot cleaner. The weird way I handle the ID is changed, we just pass the block ID and use the sign's string from there. This packet seems to be needed, trying to open the GUI any other way crashes the game and registering it with a GUI ID also doesn't work presumably because it has no container, and usually Forge GUI IDs have an associated container. Warped wart block uses a meta on the nether wart block Remove nether wart block uncrafting recipe. I didn't even notice that, that's not vanilla at all. I wonder when that happened, oops!
New signs use an ItemBlock for their sign instead of a different item ID. New fences will use this ID instead of wasting a separate ID for each one.

Still needs worldgen code, fungus currently does not grow as well. Nylium does not have proper bone meal support yet. Nether roots don't use the right potted icon yet.

+++ src/main/java/ganymedes01/etfuturum/network/WoodSignOpenMessage.java
@@ -14,7 +14,7 @@ public class WoodSignOpenMessage implements IMessage {
 	public WoodSignOpenMessage() {
 	}
 
-	public WoodSignOpenMessage(TileEntityWoodSign tileentitysign, byte i) {
+	public WoodSignOpenMessage(TileEntityWoodSign tileentitysign, int i) {
 		tileX = tileentitysign.xCoord;
 		tileY = tileentitysign.yCoord;
 		tileZ = tileentitysign.zCoord;
@@ -26,7 +26,7 @@ public class WoodSignOpenMessage implements IMessage {
 		this.tileX = buf.readInt();
 		this.tileY = buf.readInt();
 		this.tileZ = buf.readInt();
-		this.id = buf.readByte();
+		this.id = buf.readInt();
 	}
 
 	@Override
@@ -34,7 +34,7 @@ public class WoodSignOpenMessage implements IMessage {
 		buf.writeInt(this.tileX);
 		buf.writeInt(this.tileY);
 		buf.writeInt(this.tileZ);
-		buf.writeByte(this.id);
+		buf.writeInt(this.id);
 	}
 
 }
